Getting redundant errors while adding inheritance and proper encapsulation

+++ PokeBattle/Weakness.h
@@ -5,6 +5,6 @@ struct Weakness
 {
 
 	std::string Name;
-	double Multiplier;
+	int Multiplier;
 
 };
